pyproject.toml
[tool.poetry]
name = "django-and-docker"
version = "0.1.0"
description = ""
authors = ["radik <radikmanasyan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
